requirements.txt
discord.py
dislash.py
python-dotenv
asyncio
jishaku
